WIP #has_disk? can still find a disk after vMotion before detaching
This is a working implementation of a find_disk call that can still
find the disk even after it has been vMotioned. This is accomplished by
setting a custom field on the datacenter level that maps the disk_cid
to the vm_cid, so as long as the VM exists and has a persistent disk property
that matches the disk_cid, the disk can still be found.

This could still use some better organization of logic.

[#109647258](https://www.pivotaltracker.com/story/show/109647258)