Adds #pick_cluster to the ClusterPicker
This #pick_cluster method will first select clusters with the least
migration burden (smallest amount of total disk size to migrate) or no
disks to migrate. If there is a tiebreaker betweeen clusters, then the
ClusterPicker will select a suitable cluster with the greatest amount of
available resources.

This commit also factors in the case where the set of ephemeral
datastores can overlap with the set of persistent datastores.

This commit also fixes a bug in the DatastorePicker where headroom was
not taken into account for the #can_accomodate_disks? method.

[#117369019](https://www.pivotaltracker.com/story/show/117369019)

Signed-off-by: Felix Riegger <felix.riegger@sap.com>